requirements.txt
pygame>=2.5.2
python-chess>=1.999
requests>=2.31.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
Pillow>=10.1.0
gTTS>=2.3.2
berserk>=0.13.0  # Python client for the lichess API
numpy>=1.26.0
